requirements.txt
face_recognition >=1.3.0
Flask >3.0.0,<3.0.3	
flask_cors==5.0
numpy >=1.24.4
opencv_python >4.10.0,<4.10.4
pandas==2.0
Pillow >=10.0.0,<10.4.0
ultralytics >8.2.1,<8.2.101
